backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic[email]==2.7.1
motor==3.4.0
pymongo==4.7.2
PyJWT==2.8.0
bcrypt==4.1.3
anthropic==0.25.0
python-dotenv==1.0.1
